requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
PyMySQL==1.1.0
bcrypt==4.0.1
python-dateutil==2.8.2
requests==2.31.0
paystackapi==2.1.0
PyJWT==2.8.0
